+++ R/TruncatedStudentsT.R
@@ -1,4 +1,8 @@
-TruncatedStudentsT <- function(df, location = 0, scale = 1, left = -Inf, right = Inf) {
+TruncatedStudentsT <- function(df = numeric(), location = NULL, scale = NULL, left = NULL, right = NULL) {
+  if (is.null(location)) location <- rep.int(0, length(df))
+  if (is.null(scale))    scale    <- rep.int(1, length(df))
+  if (is.null(left))     left     <- rep.int(-Inf, length(df))
+  if (is.null(right))    right    <- rep.int(Inf, length(df))
   n <- c(length(df), length(location), length(scale), length(left), length(right))
   stopifnot("parameter lengths do not match (only scalars are allowed to be recycled)" = all(n %in% c(1L, max(n))))
   d <- data.frame(df = df, location = location, scale = scale, left = left, right = right)
@@ -9,18 +13,6 @@ TruncatedStudentsT <- function(df, location = 0, scale = 1, left = -Inf, right =
 mean.TruncatedStudentsT <- function(x, ...) {
   m <- ett(df = x$df, location = x$location, scale = x$scale, left = x$left, right = x$right)
   setNames(m, names(x))
-}
-
-variance.TruncatedStudentsT <- function(x, ...) {
-  stop("not yet implemented")
-}
-
-skewness.TruncatedStudentsT <- function(x, ...) {
-  stop("not yet implemented")
-}
-
-kurtosis.TruncatedStudentsT <- function(x, ...) {
-  stop("not yet implemented")
 }
 
 random.TruncatedStudentsT <- function(x, n = 1L, drop = TRUE, ...) {
@@ -72,4 +64,42 @@ is_discrete.TruncatedStudentsT <- function(d, ...) {
 
 is_continuous.TruncatedStudentsT <- function(d, ...) {
   setNames(rep.int(TRUE, length(d)), names(d))
+}
+
+score.TruncatedStudentsT <- function(d, x, which = NULL, drop = TRUE, ...) {
+  if (is.null(which)) which <- c("df", "mu", "sigma")
+  which <- gsub("scale", "sigma", which, fixed = TRUE)
+  which <- gsub("location", "mu", which, fixed = TRUE)
+  s <- stt(x, location = d$location, scale = d$scale, df = d$df, left = d$left, right = d$right, which = which, drop = drop)
+  if (!is.null(nam <- names(d))) {
+    if (is.null(dim(s))) {
+      names(s) <- nam
+    } else {
+      rownames(s) <- nam
+    }
+  }
+  if (!is.null(dim(s))) {
+    colnames(s) <- gsub("sigma", "scale", colnames(s), fixed = TRUE)
+    colnames(s) <- gsub("mu", "location", colnames(s), fixed = TRUE)
+  }
+  return(s)
+}
+
+hessian.TruncatedStudentsT <- function(d, x, which = NULL, drop = TRUE, expected = FALSE, ...) {
+  if (is.null(which)) which <- c("df", "mu:df", "sigma:df", "df:mu", "mu", "sigma:mu", "df:sigma", "mu:sigma", "sigma")
+  which <- gsub("scale", "sigma", which, fixed = TRUE)
+  which <- gsub("location", "mu", which, fixed = TRUE)
+  h <- htt(x, location = d$location, scale = d$scale, df = d$df, left = d$left, right = d$right, which = which, drop = drop, expected = expected)
+  if (!is.null(nam <- names(d))) {
+    if (is.null(dim(h))) {
+      names(h) <- nam
+    } else {
+      rownames(h) <- nam
+    }
+  }
+  if (!is.null(dim(h))) {
+    colnames(h) <- gsub("sigma", "scale", colnames(h), fixed = TRUE)
+    colnames(h) <- gsub("mu", "location", colnames(h), fixed = TRUE)
+  }
+  return(h)
 }